backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
whitenoise==6.7.0
pytz==2024.1
reportlab==4.0.7
gunicorn==21.2.0

# Testing dependencies
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.2.0
responses==0.23.3

